moved new button component into /components/Button and made the component more flexible by removing explicit icon mapping
