+++ app/components/Button/Button.scss
@@ -1,48 +1,61 @@
 @import '../../styles/variables';
 
-.button {
-    padding: 6px 12px;
-    border-width: 1px;
-    border-style: solid;
-    font-size: 1em;
-    color: $button-text;
-    background-color: $default-button;
-    cursor: pointer;
+.buttonRedesign {
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  font-size: 16px;
+  font-weight: 100;
+  padding: 10px;
+}
 
-    + .button, + a .button {
-        margin-left: 10px;
-    }
+.darkButton {
+  background: $primary-button;
+  color: #fff;
+}
 
-    &.primary {
-        background-color: $default-button;
-        border-color: $default-button;
-        &:hover:not([disabled]) {
-            background-color: $default-button-hover;
-            border-color: $default-button-hover;
-        }
-    }
+.buttonRedesign:hover {
+  background: linear-gradient(to right, #a866ee, #ee6d66);
+}
 
-    &.secondary {
-        background-color: $dark-button;
-        border-color: $dark-button;
-        &:hover:not([disabled]) {
-            background-color: $dark-button-hover;
-            border-color: $dark-button-hover;
-        }
-    }
+.buttonRedesign span {
+  padding-right: 25px;
+  margin: auto;
+}
 
-    &.cancel {
-        background-color: transparent;
-        border-color: #999;
-        color: #000;
-        &:hover:not([disabled]) {
-            background-color: transparent;
-            border-color: #999;
-        }
-    }
+.icon {
+  height: 25px;
+  width: 25px;
+  padding-left: 10px;
+  align-self: flex-start;
+}
 
-    &[disabled] {
-        opacity: .6;
-        cursor: default;
-    }
+.darkButton path {
+  fill: #fff;
+}
+
+.lightButton {
+  border: solid;
+  border-width: 2px;
+  border-color: $primary-button;
+  background: #fff;
+  color: $primary-button;
+}
+
+.lightButton:hover {
+  border: double 2px transparent;
+  background-image: linear-gradient(to right, #ee6d66, #a866ee),
+    radial-gradient(circle at top left, #ee6d66, #a866ee);
+  background-origin: border-box;
+  background-clip: content-box, border-box;
+  border-color: transparent;
+  border-color: transparent;
+  color: white;
+  path {
+    fill: #fff;
+  }
 }
