--- app/components/Button/Redesign/index.js
@@ -1,1 +1,0 @@
-export { default } from './Button'
